Added and used helper functions for auth and permissions
- Split permisson enum into its own file

+++ websites/admin/public/categories.php
@@ -1,8 +1,13 @@
 <?php
 
+use Trulyao\Eds\Auth;
 use Trulyao\Eds\Models\Category;
+use Trulyao\Eds\Models\Permission;
 
 require_once __DIR__ . "/../src/prelude.php";
+
+require_auth();
+ensure_user_can(Permission::Read);
 
 define("PAGE_SIZE", 25);
 
@@ -37,15 +42,19 @@ render_header("Categories");
                   <?php echo $category->name ?>
                 </th>
                 <td class="actions">
-                  <a href="/manage/category.php?id=<?php echo $category->id; ?>" class="action">
-                    <i class="ti ti-edit"></i>
-                    <span>edit</span>
-                  </a>
+                  <?php if (Auth::getUser()?->can(Permission::Write)) : ?>
+                    <a href="/manage/category.php?id=<?php echo $category->id; ?>" class="action">
+                      <i class="ti ti-edit"></i>
+                      <span>edit</span>
+                    </a>
+                  <?php endif; ?>
 
-                  <a href="/manage/category.php?id=<?php echo $category->id; ?>&action=delete" class="action delete" data-confirm="Are you sure you want to delete this category? THIS ACTION CANNOT BE REVERSED.">
-                    <i class="ti ti-trash"></i>
-                    <span>delete</span>
-                  </a>
+                  <?php if (Auth::getUser()?->can(Permission::Delete)) : ?>
+                    <a href="/manage/category.php?id=<?php echo $category->id; ?>&action=delete" class="action delete" data-confirm="Are you sure you want to delete this category? THIS ACTION CANNOT BE REVERSED.">
+                      <i class="ti ti-trash"></i>
+                      <span>delete</span>
+                    </a>
+                  <?php endif; ?>
                 </td>
               </tr>
             <?php endforeach; ?>
